refactor: move legacy tx history-related structs to same file
