--- rust/src/state/constants.rs
@@ -1,39 +1,0 @@
-use serde::{Deserialize, Serialize};
-use spdk_wallet::bitcoin::OutPoint;
-use spdk_wallet::bitcoin::{absolute::Height, Amount, BlockHash, Txid};
-use spdk_wallet::client::Recipient;
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
-pub enum RecordedTransaction {
-    Incoming(RecordedTransactionIncoming),
-    Outgoing(RecordedTransactionOutgoing),
-    UnknownOutgoing(RecordedTransactionUnknownOutgoing),
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
-pub struct RecordedTransactionIncoming {
-    pub txid: Txid,
-    pub amount: Amount,
-    pub confirmation_height: Option<Height>,
-    pub confirmation_blockhash: Option<BlockHash>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
-pub struct RecordedTransactionOutgoing {
-    pub txid: Txid,
-    pub spent_outpoints: Vec<OutPoint>,
-    pub recipients: Vec<Recipient>,
-    pub confirmation_height: Option<Height>,
-    pub confirmation_blockhash: Option<BlockHash>,
-    pub change: Amount,
-    #[serde(default)]
-    pub fee: Amount,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
-pub struct RecordedTransactionUnknownOutgoing {
-    pub spent_outpoints: Vec<OutPoint>,
-    pub amount: Amount,
-    pub confirmation_height: Height,
-    pub confirmation_blockhash: BlockHash,
-}
