+++ rust/src/state/mod.rs
@@ -1,4 +1,3 @@
-pub mod constants;
 mod updater;
 
 pub use updater::StateUpdater;
